requirements.txt
django-filter
celery
boto3==1.9.133
botocore==1.12.133
Django==2.2
django-debug-toolbar==1.8
django-multiselectfield==0.1.8
django-storages==1.7.1
docutils==0.14
jmespath==0.9.4
mysqlclient==1.4.2.post1
Pillow==6.0.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
raven==6.10.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
